fix: SurveyForm 緊急度 Select の自動クリアを修正（コードレビュー反映）
- urgencyOptions の先頭空 option を廃し allowEmpty+placeholder に（InterviewForm と統一）。
  未ロード時 options.length=0 で Select 自動フォールバックが発火せず、編集時の緊急度が
  null に上書きされる BLOCKER を解消
- GET 詳細の urgency_id を ::int キャストし型(number)と実体を一致（bigint 文字列化のズレ解消）

+++ apps/web/app/api/v1/answers/[id]/route.ts
@@ -18,7 +18,7 @@ export const GET = withActiveUser(async (_req, claims, { params }: Ctx) => {
              a.interview_at     as "interviewAt",
              a.interview_memo   as "interviewMemo",
              a.next_action      as "nextAction",
-             a.urgency_id       as "urgencyId",
+             a.urgency_id::int  as "urgencyId",
              a.evaluation,
              u.name             as "respondentName",
              s.id               as "surveyId",
